refactor: fetch 함수 분리
