+++ src/pages/CreatePostPage/CreatePostPage.tsx
@@ -7,8 +7,7 @@ import Calendar from '../../assets/images/calendar.png';
 import Add from '../../assets/images/add.png';
 import useInputs from '../../hooks/useInputs';
 import { CreatePost } from '../../types/post';
-
-const URL = process.env.API_URL;
+import { fetchPost } from '../../api/posts/posts';
 
 const inputStyles = {
   borderRadius: '5px',
@@ -17,16 +16,6 @@ const inputStyles = {
   border: 0,
   backgroundColor: '#EEEEEE',
   boxShadow: '0px 4px 4px gray',
-};
-
-const fetchPost = async (postForm: CreatePost) => {
-  const response = await fetch(`${URL}/posts`, {
-    method: 'POST',
-    credentials: 'include',
-    body: JSON.stringify(postForm),
-  });
-
-  return response.status;
 };
 
 const CreatePostPage = () => {
