+++ src/pages/LandingPage/LandingPage.tsx
@@ -5,14 +5,7 @@ import MainPost from '../../components/MainPost';
 import PostImage from '../../components/PostImage';
 import Carousel from '../../components/Carousel';
 import { Post } from '../../types/post';
-
-const URL = process.env.API_URL;
-
-const fetchPostsData = async (): Promise<Post[]> => {
-  const response = await fetch(`${URL}/posts`);
-  const result = (await response.json()) as Post[];
-  return result;
-};
+import { fetchPostsData } from '../../api/posts/posts';
 
 const LandingPage = () => {
   const { data } = useQuery({
